Updated the Typing System
A little update which allows for a tutorial of sorts.
The first 'level' allows the player to type without worries, there's no player movement or approaching fog.

TODO:
The second 'level' should allow jumping, but it is not working as of yet.
The third 'level' will enable player movement.
The 'forth' will set the fog in motion.